Move modals to modals package, adjust font size in modals
