+++ components/board-component.tsx
@@ -2,11 +2,11 @@ import {WorkflowList, WorkflowListType} from "utils/models";
 import {Draggable, Droppable} from "react-beautiful-dnd";
 import ListComponent from "components/list-component";
 import React, {useEffect, useState} from "react";
-import CreateWorkflowListModal from "components/create-workflowlist-modal";
-import ModifyWorkflowListModal from "components/modify-workflowlist-modal";
+import CreateWorkflowListModal from "components/modals/create-workflowlist-modal";
+import ModifyWorkflowListModal from "components/modals/modify-workflowlist-modal";
 import ItemComponent from "components/item-component";
 import {getDroppableStyle} from "utils/style-elements";
-import MoveWorkflowListModal from "components/move-workflowlist-modal";
+import MoveWorkflowListModal from "components/modals/move-workflowlist-modal";
 import DropButton from "components/drop-button";
 import ButtonsMenu from "components/buttons-menu";
 import {formatDate} from "utils/date-util";
