+++ components/modals/create-workflowlist-modal.tsx
@@ -36,7 +36,7 @@ const CreateWorkflowListModal = ({
     return (
         <div style={showHideStyle}
              className="fixed z-10 inset-0 overflow-y-auto">
-            <div className="flex items-end justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0">
+            <div className="flex items-end justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0 text-sm">
                 <div className="fixed inset-0 transition-opacity" aria-hidden="true">
                     <div className="absolute inset-0 bg-gray-500 opacity-75"/>
                 </div>
